requirements.txt
attrs==22.2.0
aws-cdk-lib==2.71.0
aws-cdk.asset-awscli-v1==2.2.142
aws-cdk.asset-kubectl-v20==2.1.1
aws-cdk.asset-node-proxy-agent-v5==2.0.118
aws-cdk.cloud-assembly-schema==1.198.1
aws-cdk.cx-api==1.198.1
aws-cdk.region-info==1.198.1
boto3==1.26.114
botocore==1.29.114
cattrs==22.2.0
constructs==10.1.314
exceptiongroup==1.1.1
importlib-resources==5.12.0
iniconfig==2.0.0
jmespath==1.0.1
jsii==1.80.0
packaging==23.1
pluggy==1.0.0
publication==0.0.3
py==1.11.0
pytest==6.2.5
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
toml==0.10.2
typeguard==2.13.3
typing_extensions==4.5.0
urllib3==1.26.15
uuid==1.30
zipp==3.15.0
